remake init upload
