--- patch_instructions.json
@@ -1,92 +1,0 @@
-{
-    "10.0.0": {
-        "release": {
-            "0.0.1" : [
-                "url_add_mod|Magic_Wesley|Atlantica|3.0.6",
-                "url_add_mod|Magic_Wesley|Desolation|2.0.6",
-                "url_add_mod|Magic_Wesley|Oldred|2.0.9",
-                "url_add_mod|Magic_Wesley|WesleysInteriors|1.1.17",
-                "url_add_mod|sfDesat|Aquatis|2.2.3",
-                "url_add_mod|sfDesat|Celest|1.1.4",
-                "url_add_mod|sfDesat|Celestial_Tint|1.5.1",
-                "url_add_mod|stayhydrated|SillycoreVids|1.1.1",
-                "url_add_mod|TestAccount666|ShipWindows|1.3.6",
-                "url_add_mod|Verity|TooManySuits|1.1.0",
-                "url_add_mod|xilophor|LethalNetworkAPI|2.2.0",
-                "url_add_mod|Yorimor|CustomStoryLogs|1.4.4",
-                "delete_mod|SelfKick|SelfKick",
-                "update_vers|0.0.1"
-            ],
-            "0.0.2": [
-                "url_add_mod|DiFFoZ|BepInEx_Faster_Load_AssetBundles_Patcher|0.6.2",
-                "url_add_mod|qwbarch|BarchLib|1.0.2",
-                "url_add_mod|qwbarch|Mirage|1.7.1",
-                "url_add_mod|TheWeavers|NightmareFreddy|1.1.2",
-                "url_add_mod|XuXiaolan|TheGiantSpecimens|2.1.5",
-                "download_override|18m19IWHFzkhSRRQwI7Y8-9cfdhGbqetj|https://www.dropbox.com/scl/fi/ywhw23zd4gh7o8sfewsp0/PBLC-v10-patch-2.zip?rlkey=77a0wvurndi7c55efqoz9kt7w&st=gkrhmpmd&dl=1",
-                "update_vers|0.0.2"
-            ],
-            "0.0.3": [
-                "url_add_mod|DiFFoZ|BepInEx_Faster_Load_AssetBundles_Patcher|0.6.3",
-                "url_add_mod|Magic_Wesley|WesleysInteriors|1.1.18",
-                "update_vers|0.0.3"
-            ],
-            "0.0.4": [
-                "url_add_mod|BunyaPineTree|ModelReplacementAPI|2.4.6",
-                "url_add_mod|ButteryStancakes|AccurateStaminaDisplay|2.1.3",
-                "url_add_mod|DiFFoZ|BepInEx_Faster_Load_AssetBundles_Patcher|0.6.4",
-                "url_add_mod|Generic_GMD|Generic_Moons|5.0.0",
-                "url_add_mod|Magic_Wesley|WesleysInteriors|1.1.19",
-                "url_add_mod|MasterAli2|TeleportDecline|0.1.0",
-                "url_add_mod|NomnomAB|RollingGiant|2.6.0",
-                "url_add_mod|Quixler|DeadAndBored|1.2.0",
-                "url_add_mod|qwbarch|Mirage|1.7.3",
-                "url_add_mod|Tolian|PsychSanctum|1.0.0",
-                "download_override|1QomCWAggPpuAqy6szqlz-tQlIASuwIvq|https://www.dropbox.com/scl/fi/rpzunicx786dsiiojf4dj/PBLC-v10-patch-4.zip?rlkey=k2eb2oziz3xv862uozfczz83o&st=hkg9wvza&dl=1",
-                "update_vers|0.0.4"
-            ],
-            "0.0.5": [
-                "download_override|1axp_QRYunzEwdNihu9URGg0qiyXxf0I5|https://www.dropbox.com/scl/fi/v2mm52w1xchjdrsmgsomh/PBLC-v10-patch-5.zip?rlkey=ukpg9ys8akq2roq6hieaz2163&st=9ctd875c&dl=1",
-                "update_vers|0.0.5"
-            ],
-            "0.0.6": [
-                "url_add_mod|FlipMods|ObjectVolumeController|1.1.0",
-                "url_add_mod|JackdeDestroyer|More_Company_TV_Channels|1.3.0",
-                "url_add_mod|mrov|MrovLib|0.0.10",
-                "url_add_mod|skidz|PoolRooms|0.1.22",
-                "delete_mod|Kesomannen|KeepItDown",
-                "download_override|1k9k-X9allfbNijQVq04N2HVP-IPbRMJc|https://www.dropbox.com/scl/fi/rrdkilhr9gtf1l80ytehk/PBLC-v10-patch-6.zip?rlkey=de6fnwxomq6col56et7mz1c2m&st=emba7xqp&dl=1",
-                "update_vers|0.0.6"
-            ],
-            "0.0.7": [
-                "url_add_mod|Generic_GMD|Generic_Moons|5.1.0",
-                "url_add_mod|necrowing|StopBeesOpeningDoors|1.1.0",
-                "delete_mod|MasterAli2|TeleportDecline",
-                "download_override|1IzjgHMn2oxBpserYCCLZGJVgmHG84X4n|https://www.dropbox.com/scl/fi/t2lpczb0ga5i880x4ak68/PBLC-v10-patch-7.zip?rlkey=u8kv89pk9yb9pkuybmycnuxq7&st=ib4nf5pb&dl=1",
-                "update_vers|0.0.7"
-            ],
-            "0.0.8": [
-                "url_add_mod|Gemumoddo|LethalEmotesAPI|1.12.8",
-                "url_add_mod|malco|Lategame_Upgrades|3.8.1",
-                "url_add_mod|TestAccount666|ShipWindows|1.3.7",
-                "url_add_mod|WhiteSpike|Interactive_Terminal_API|1.1.2",
-                "url_add_mod|Zaggy1024|OpenBodyCams|2.2.1",
-                "url_add_mod|ButteryStancakes|MeleeFixes|1.2.0",
-                "delete_mod|MasterAli2|TeleportDecline",
-                "delete_mod|RosiePies|Sector0_Interior",
-                "download_override|1SXj_2Ih-_awf1SMXE2nqvDxwOkHsHn9c|https://www.dropbox.com/scl/fi/7dtcxuqmxd392qov0bz08/PBLC-v10-patch-8.zip?rlkey=eiwibf6271sgug1xtwnmo4vyh&st=n4npre1r&dl=1",
-                "update_vers|0.0.8"
-            ],
-            "0.0.9": [
-                "url_add_mod|RosiePies|Rosies_Moons|2.0.2",
-                "url_add_mod|RosiePies|Sector0_Interior|2.0.1",
-                "update_vers|0.0.9"
-            ],
-            "0.1.0": [
-                "delete_mod|malco|Lategame_Upgrades",
-                "download_override|1NlCkdTG1GOE90bh9vMKeqf3YajJyh9lU|https://www.dropbox.com/scl/fi/bnboeyg5t739qgfqiso8g/PBLV-v10-patch-10.zip?rlkey=0mms0qdqlik0qmiphk6hjt5kp&st=0n1ptw2y&dl=1",
-                "update_vers|0.1.0"
-            ]
-        }
-    }
-}
