--- update_files/10.0.0.json
@@ -1,114 +1,0 @@
-{
-    "instructions": [
-        "clean_install",
-        "download_bepinex",
-        "url_add_mod|AinaVT|LethalConfig|1.4.2",
-        "url_add_mod|AinaVT|YippeeScrap|1.0.7",
-        "url_add_mod|akronman1|Niko|0.2.1",
-        "url_add_mod|AllToasters|SpectateEnemies|2.4.1",
-        "url_add_mod|anormaltwig|LateCompany|1.0.13",
-        "url_add_mod|Atomicz_A|ScrollReverser|1.0.1",
-        "url_add_mod|AudioKnight|StarlancerAIFix|3.6.0",
-        "url_add_mod|Beauver|ExtraTerminalCommands|1.5.0",
-        "url_add_mod|Bibendi|AEIOUCompany|1.3.6",
-        "url_add_mod|bigmcnugget|VanillaContentExpansion|0.1.8",
-        "url_add_mod|Blorb|WeatherMultipliers|1.1.0",
-        "url_add_mod|Bobbie|NAudio|2.2.2",
-        "url_add_mod|Bobbie|UniTask|2.5.0",
-        "url_add_mod|BunyaPineTree|ModelReplacementAPI|2.4.5",
-        "url_add_mod|ButteryStancakes|AccurateStaminaDisplay|2.1.2",
-        "url_add_mod|ButteryStancakes|SnowyHolidayDropship|1.0.4",
-        "url_add_mod|Clementinise|CustomSounds|2.3.2",
-        "url_add_mod|coderCleric|Poltergeist|1.1.0",
-        "url_add_mod|Coppertiel|MaskedEnemyOverhaulFork|3.3.0",
-        "url_add_mod|darmuh|ghostCodes|2.0.5",
-        "url_add_mod|DiFFoZ|BepInEx_Faster_Load_AssetBundles_Patcher|0.5.0",
-        "url_add_mod|EliteMasterEric|Coroner|1.6.2",
-        "url_add_mod|EliteMasterEric|SlimeTamingFix|1.0.2",
-        "url_add_mod|Evaisa|FixPluginTypesSerialization|1.1.1",
-        "url_add_mod|Evaisa|HookGenPatcher|0.0.5",
-        "url_add_mod|Evaisa|LethalLib|0.16.0",
-        "url_add_mod|fivetoofive|DetailedScan|1.2.2",
-        "url_add_mod|FlipMods|TooManyEmotes|2.1.19",
-        "url_add_mod|FlipMods|TooManyEmotesScrap|1.0.5",
-        "url_add_mod|fumiko|CullFactory|1.1.3",
-        "url_add_mod|Hardy|LCMaxSoundsFix|1.2.0",
-        "url_add_mod|HGG|JigglePhysicsPlugin|1.1.21",
-        "url_add_mod|IAmBatby|LethalLevelLoader|1.2.4",
-        "url_add_mod|IAmBatby|LethalToolbox|1.0.2",
-        "url_add_mod|IntegrityChaos|Diversity|2.1.0",
-        "url_add_mod|IntegrityChaos|GraphicsAPI|1.0.0",
-        "url_add_mod|IntegrityChaos|LCCutscene|1.0.0",
-        "url_add_mod|jaspercreations|Scopophobia|1.1.2",
-        "url_add_mod|Jordo|NeedyCats|1.2.1",
-        "url_add_mod|Kesomannen|KeepItDown|1.4.0",
-        "url_add_mod|Kittenji|Groan_Tube_Scrap|1.0.1",
-        "url_add_mod|KlutzyBubbles|BetterEmotes|1.5.1",
-        "url_add_mod|LethalMatt|Bozoros|1.0.1",
-        "url_add_mod|LethalMatt|Outpost_31|2.0.0",
-        "url_add_mod|linkoid|DissonanceLagFix|1.0.0",
-        "url_add_mod|Lordfirespeed|OdinSerializer|2022.11.9",
-        "url_add_mod|Magic_Wesley|Acidir|5.0.6",
-        "url_add_mod|Magic_Wesley|Asteroid13|3.0.9",
-        "url_add_mod|Magic_Wesley|Atlantica|3.0.5",
-        "url_add_mod|Magic_Wesley|Desolation|2.0.5",
-        "url_add_mod|Magic_Wesley|Etern|4.0.7",
-        "url_add_mod|Magic_Wesley|Infernis|4.0.5",
-        "url_add_mod|Magic_Wesley|Oldred|2.0.8",
-        "url_add_mod|Magic_Wesley|WesleysInteriors|1.1.16",
-        "url_add_mod|Major_And_Skiz|MoreInteriors|2.5.0",
-        "url_add_mod|MarshMan|DeezNuts|1.0.0",
-        "url_add_mod|MaxWasUnavailable|LethalModDataLib|1.2.2",
-        "url_add_mod|MegaPiggy|EnumUtils|1.0.3",
-        "url_add_mod|MikeMediaStudios|MikesTweaks|2.0.2",
-        "url_add_mod|MrHydralisk|EnhancedRadarBooster|1.7.1",
-        "url_add_mod|mrov|MrovLib|0.0.9",
-        "url_add_mod|mrov|TerminalFormatter|0.2.13",
-        "url_add_mod|no00ob|LCSoundTool|1.5.1",
-        "url_add_mod|NotAtomicBomb|TerminalApi|1.5.4",
-        "url_add_mod|Ozone|Runtime_Netcode_Patcher|0.2.5",
-        "url_add_mod|Pooble|LCBetterSaves|1.7.3",
-        "url_add_mod|qwbarch|BarchLib|1.0.0",
-        "url_add_mod|qwbarch|Mirage|1.7.0",
-        "url_add_mod|Rattenbonkers|TVLoader|1.1.1",
-        "url_add_mod|redroacter|Necoarc_suits|0.2.0",
-        "url_add_mod|RosiePies|Rosies_Moons|2.0.0",
-        "url_add_mod|RosiePies|Sector0_Interior|1.1.4",
-        "url_add_mod|Rune580|LethalCompany_InputUtils|0.7.7",
-        "url_add_mod|Scintesto|Monster_Plushies|1.2.9",
-        "url_add_mod|Scintesto|Scavenger_Plushie|1.0.9",
-        "url_add_mod|Sconeys|abandonedcompanyassets|0.15.79",
-        "url_add_mod|scoopy|Scoopys_Variety_Mod|1.2.0",
-        "url_add_mod|SelfKick|SelfKick|1.2.0",
-        "url_add_mod|sfDesat|Aquatis|2.2.2",
-        "url_add_mod|sfDesat|Celest|1.1.3",
-        "url_add_mod|sfDesat|Celestial_Tint|1.4.6",
-        "url_add_mod|sfDesat|ViewExtension|1.2.2",
-        "url_add_mod|ShaosilGaming|GeneralImprovements|1.2.8",
-        "url_add_mod|Sigurd|CSync|5.0.1",
-        "url_add_mod|SimonTendo|LCGoldScrapMod|1.7.0",
-        "url_add_mod|Sparble|FacelessStalker|1.1.5",
-        "url_add_mod|stayhydrated|SillycoreVids|1.0.0",
-        "url_add_mod|sunnobunno|YippeeMod|1.2.4",
-        "url_add_mod|SweetOnion|LethalSnap|1.5.7",
-        "url_add_mod|SylviBlossom|TerminalConflictFix|1.2.2",
-        "url_add_mod|SZAKI|EnemyFix|1.0.0",
-        "url_add_mod|taffyko|BetterSprayPaint|2.0.4",
-        "url_add_mod|TestAccount666|ShipWindows|1.2.3",
-        "url_add_mod|TheWeavers|LethalBestiary|1.2.1",
-        "url_add_mod|TheWeavers|NightmareFoxy|1.1.10",
-        "url_add_mod|TheWeavers|NightmareFreddy|1.0.1",
-        "url_add_mod|tinyhoot|ShipLoot|1.1.0",
-        "url_add_mod|Verity|TooManySuits|1.0.5",
-        "url_add_mod|x753|More_Suits|1.4.3",
-        "url_add_mod|xilophor|LethalNetworkAPI|2.1.7",
-        "url_add_mod|XuXiaolan|ImmersiveScrap|1.1.0",
-        "url_add_mod|XuXiaolan|TheGiantSpecimens|2.1.2",
-        "url_add_mod|Yorimor|CustomStoryLogs|1.3.3",
-        "url_add_mod|Zaggy1024|OpenBodyCams|2.2.0",
-        "url_add_mod|Zaggy1024|PathfindingLagFix|1.2.1",
-        "url_add_mod|ZG|ScalingStartCredits|1.0.6",
-        "download_override|1WS208_gZK6458tmr4o0udzG9fn4i-3az|https://www.dropbox.com/scl/fi/r13bvndfjmoyfchfrawtj/PBLC-v10-Overrides.zip?rlkey=bua48fxciabapennkn2wn2116&st=hz3knkwl&dl=1",
-        "update_vers|0.0.0|10.0.0"
-    ]
-}
